Moved Trusts controller tests to integration, created test class for unit tests
