+++ TramsDataApi.Test/Controllers/TrustsControllerTests.cs
@@ -1,0 +1,9 @@
+using Xunit;
+
+namespace TramsDataApi.Test.Controllers
+{
+    public class TrustsControllerTests
+    {
+        
+    }
+}
